OSPF: Modernize some code.
Make a few one-byte fetches use GET_U_1().

In ospf_print_tos_metrics() lose an excess ND_TCHECK_SIZE() and make the
function void.

Make ospf_decode_lls() void -- the only case when it returned a non-zero
value was if the LLS block is invalid, and it already signals that with a
message. It is a separate matter if the invalid packet is also truncated,
and since there is no proper check for that, just return without printing
a misleading message.

Switch ospf_print() to nd_trunc_longjmp().